Moved sound assets to separated subspec
